Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/parsing/font-variant-east-asian-invalid-expected.txt


PASS e.style['font-variant-east-asian'] = "normal ruby" should not set the property value
FAIL e.style['font-variant-east-asian'] = "jis78 jis83" should not set the property value assert_equals: expected "" but got "jis83"
FAIL e.style['font-variant-east-asian'] = "full-width proportional-width" should not set the property value assert_equals: expected "" but got "proportional-width"
